refactor: move friend request model
